requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.8.0
    # via -r requirements.in
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via uvicorn
colorama==0.4.5
    # via click
fastapi==0.80.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
fastapi-cache2==0.1.9
    # via -r requirements.in
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   rfc3986
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
pendulum==2.1.2
    # via fastapi-cache2
pydantic==1.9.2
    # via fastapi
python-dateutil==2.8.2
    # via pendulum
pytzdata==2020.1
    # via pendulum
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via python-dateutil
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.19.1
    # via fastapi
typing-extensions==4.3.0
    # via
    #   pydantic
    #   starlette
uvicorn==0.18.3
    # via
    #   -r requirements.in
    #   fastapi-cache2
